Refactored the "move coordinate" function
The issue I had at first was that I needed to move the coordinate in the opposite direction when spawning the snake. But it can be confusing having the function to do that, so it occurred to me that I could just use a negative offset. Even though the size_t type is unsigned, the wraparound behavion of unsigned integers is going to make the position end up on the intended place anyways.

Also that function is going to be used later but with positive offsets, so I wont't need to create a new function for moving the coordinate.