Check for the glibc version before using the `__fpurge()` function
If the check fails, then we fall back to the regular `tcflush()` function. I prefer using `__fpurge()` because it clears stdin immediately, while `tchflush()` appears to have a little delay before that.

However, `__fpurge()` is not always available. On the other hand, `tcflush()` should be available on any C implementation on Linux. Thus the need to check for glibc and its version, to determine which function to use. The man page says `__fpurge()` is available from glibc 2.1.95 onwards.