Changed when the CodeQL action runs
Now it is on pushes or pull requests of any branches, and also on every Thrursday at 15:30.

I also changed the build so the project is compiled with debug symbols. That should help in the sanitizers output an error during runtime (it is going to display which parts of the source code caused the error).

And I have also enabled verbose output for GCC.