Updated version number to 1.0.4
I had originally intended to have on this release input with the "WASD" keys (QWERTY layout) on such way that the keys would be still on the same physical position on other keyboard layouts. But that proved to be trickier to implement than what I anticipated. Trickier, but not impossible, I am pushing it to version 1.0.5, with also the option to remap those input keys manually.