Created `key_mapper` translation unit
This is going to be used in order to convert a keyboard's scan code to a character in the current keyboard layout. This way we can map keys in a way that is independent of the keyboard layout (for example "WASD" in an English keyboard becomes "ZQSD" in a French keyboard).