Buffer for storing encoded characters
I made the buffer for storing UTF-8 characters into a more generic buffer, that I am going to use for any character encoding. I did that in order to not need to change input parsing on Windows to read wide chars instead of chars. I am going to convert the UTF-16 characters on Windows to multibyte, so the program can still read input with `getchar()`;